# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/accounts/like/{query}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Search for Accounts by query",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Search query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Account"
                            }
                        }
                    }
                }
            }
        },
...
        "/accounts/{id}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "description": "Display the specified Account",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Account",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Account"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Update the specified Account in storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account updated successfully",
                        "schema": {
                            "$ref": "#/definitions/Account"
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Remove the specified Account from storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
...
        "/accommodationTypes/{id}": {
            "get": {
                "tags": [
                    "AccommodationType"
                ],
                "description": "Display the specified Category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Category"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the specified Category in storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Category"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Remove the specified Category from storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Account": {
            "required": [
                "first_name",
                "last_name",
                "email"
            ],
            "properties": {
                "id": {
                    "description": "Account id",
                    "type": "integer",
                    "format": "int32"
                },
                "first_name": {
                    "description": "First name",
                    "type": "string"
                },
                "last_name": {
                    "description": "Last name",
                    "type": "string"
                },
                "activated": {
                    "description": "Activated status",
                    "type": "boolean"
                },
                "email": {
                    "description": "Email address",
                    "type": "string"
                },
                "langKey": {
                    "description": "Language key",
                    "type": "string"
                },
                "city_id": {
                    "description": "City id",
                    "type": "integer",
                    "format": "int32"
                },
                "gender": {
                    "description": "Gender",
                    "type": "string"
                },
                "image_url": {
                    "description": "Image URL",
                    "type": "string"
                },
                "dni": {
                    "description": "Document number",
                    "type": "string"
                },
                "user_id": {
                    "description": "User id",
                    "type": "integer",
                    "format": "int32"
                },
                "address": {
                    "description": "Account address",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone number",
                    "type": "string"
                },
                "birthday": {
                    "description": "Birthday",
                    "type": "string",
                    "format": "date-time"
                },
                "account_cod": {
                    "description": "Account code",
                    "type": "string"
                }
            }
        },
        "Category": {
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "description": "Category id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "Category name",
                    "type": "string"
                },
                "description": {
                    "description": "Category description",
                    "type": "string"
                },
                "category_id": {
                    "description": "Parent category id",
                    "type": "integer",
                    "format": "int32"
                },
                "color": {
                    "description": "Category color",
                    "type": "string"
                }
            }
        },
...
    }
}